# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Music Catalog API",
        "description": "API для управления исполнителями, альбомами и песнями.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Localhost API server"
        }
    ],
    "paths": {
...
        "/api/artists": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Получить список всех исполнителей",
                "operationId": "2c8aed1387cca1c65d98da7ee11ac926",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение"
                    }
                }
            },
            "post": {
                "tags": [
                    "Artists"
                ],
                "summary": "Создать нового исполнителя",
                "operationId": "83d2888bdc6fac66b9a7cc81c4c950fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Coldplay"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Исполнитель создан"
                    }
                }
            }
        },
        "/api/artists/{id}": {
            "get": {
                "tags": [
                    "Artists"
                ],
                "summary": "Получить информацию об исполнителе",
                "operationId": "d0b68c06b8bc310fbe24c6d3ea80dac3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID исполнителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация об исполнителе"
                    }
                }
            },
            "put": {
                "tags": [
                    "Artists"
                ],
                "summary": "Обновить информацию об исполнителе",
                "operationId": "5a5632ea389f9a05336f8d035f5a44c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID исполнителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Imagine Dragons"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Исполнитель обновлен"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Artists"
                ],
                "summary": "Удалить исполнителя",
                "operationId": "4f6d8bd2c8b4b738af7589b72348ac20",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID исполнителя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Исполнитель удален"
                    }
                }
            }
        },
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Получить список всех песен",
                "operationId": "0388bb28c35eec49b8b492d169ce3575",
                "responses": {
                    "200": {
                        "description": "Успешное выполнение"
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Создать новую песню",
                "operationId": "2f1970b82d91cc715578d972189f19cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Yellow"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Песня создана"
                    }
                }
            }
        },
        "/api/songs/{id}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Получить информацию о песне",
                "operationId": "4325a96361d1d3820cc0580eaae9197c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о песне"
                    }
                }
            },
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Обновить информацию о песне",
                "operationId": "452de7f8f11a121cd11e9430565632c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Fix You"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Песня обновлена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Удалить песню",
                "operationId": "829c1a52ae5013ce3ff4b9829a31485f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Песня удалена"
                    }
                }
            }
        },
        "/api/songs/{id}/add-to-album": {
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Добавить песню в альбом",
                "operationId": "5645e43aef0d6a587bff7b411c6be977",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "album_id",
                                    "track_number"
                                ],
                                "properties": {
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "track_number": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Песня добавлена в альбом"
                    }
                }
            }
        },
        "/api/songs/{id}/remove-from-album": {
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Удалить песню из альбома",
                "operationId": "2c54254fc106819009622552add57964",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID песни",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "album_id"
                                ],
                                "properties": {
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Песня удалена из альбома"
                    }
                }
            }
        }
    },
...
}